+++ src/py_bugger/cli/cli.py
@@ -1,11 +1,10 @@
-import sys
 from pathlib import Path
-import os
 
 import click
 
 from py_bugger import py_bugger
-from py_bugger.cli import cli_messages
+from py_bugger.cli import cli_utils
+from py_bugger.cli.config import pb_config
 
 
 @click.command()
@@ -32,30 +31,10 @@ from py_bugger.cli import cli_messages
     default=1,
     help="How many bugs to introduce.",
 )
-def cli(exception_type, target_dir, target_file, num_bugs):
+def cli(**kwargs):
     """Practice debugging, by intentionally introducing bugs into an existing codebase."""
-    if not exception_type:
-        click.echo(cli_messages.msg_exception_type_required)
-        sys.exit()
+    # Update pb_config using options passed through CLI call.
+    pb_config.__dict__.update(kwargs)
+    cli_utils.validate_config()
 
-    if target_dir and target_file:
-        click.echo(cli_messages.msg_target_file_dir)
-        sys.exit()
-
-    # Make sure we're passing appropriate Path objects.
-    target_dir = _set_target_dir(target_dir)
-    if target_file:
-        target_file = Path(target_file)
-
-    py_bugger.main(exception_type, target_dir, target_file, num_bugs)
-
-
-# --- Helper functions (move to a cli/utils.py module) ---
-
-
-def _set_target_dir(target_dir):
-    """Set an appropriate target directory."""
-    if target_dir:
-        return Path(target_dir)
-    else:
-        return Path(os.getcwd())
+    py_bugger.main()